Units 5, 6.1 and the midterm exam completed

+++ unidad_2/facil/CalculoResto.c
@@ -1,20 +1,16 @@
 
 #include <stdio.h>
-#include <stdlib.h>
 
-int main()
-{
-  float a b,c;
+int main() {
+
+  int a, b, c;
   printf("Introdisca un valor para encontrar el resto\n");
-  scanf("%f",&a);
+  scanf("%d", &a);
   printf("Introdisca un valor para dividir\n");
-  scanf("%f",&b);
-  c =a%b;
-
+  scanf("%d", &b);
+  c = a % b;
 
   printf("El resto es  %d.\n", (int)a);
-
-
 
   return 0;
 }
